pyproject.toml
[project]
name = "focusconnpass"
version = "0.1.0"
description = "connpass APIから直近開催イベントを取得し、ホスト側LLMが推薦を行うMCPサーバー"
requires-python = ">=3.12"
dependencies = [
    "mcp[cli]",
    "httpx",
    "pyyaml",
    "pydantic",
    "python-dotenv",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
    "ruff",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
target-version = "py312"
line-length = 88

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "N",    # pep8-naming
    "D",    # pydocstyle
    "ANN",  # flake8-annotations
    "UP",   # pyupgrade
    "B",    # flake8-bugbear
    "SIM",  # flake8-simplify
    "RUF",  # ruff-specific
]
ignore = [
    "D100",   # missing docstring in public module
    "D104",   # missing docstring in public package
]

[tool.ruff.lint.pydocstyle]
convention = "google"
